ai_engine/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI SDKs
anthropic==0.40.0
claude-agent-sdk==0.1.18
google-generativeai==0.3.2

# Database & Cache
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0

# HTTP & Tools
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Monitoring & Logging
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
